Sources/Version3/03_diaspora_image_processing/evaluation/requirements.txt
# ==============================================
# AI 기반 한인 디아스포라 기록유산 디지털화 시스템
# 이미지 처리 모듈 - Python 의존성
# ==============================================
#
# 설치 순서 (INSTALL.md 참고):
#   1. PyTorch (CUDA 12.1 빌드, 별도 설치)
#      pip install torch==2.1.2 torchvision==0.16.2 --index-url https://download.pytorch.org/whl/cu121
#   2. 본 파일의 패키지
#      pip install -r requirements.txt
#

# ──────────────────────────────────────────────
# 핵심 이미지 처리
# ──────────────────────────────────────────────
opencv-python==4.13.0.86
Pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0

# ──────────────────────────────────────────────
# 초해상도 (Real-ESRGAN)
# ──────────────────────────────────────────────
basicsr>=1.4.2
realesrgan>=0.3.0

# ──────────────────────────────────────────────
# 얼굴 복원 (GFPGAN)
# ──────────────────────────────────────────────
gfpgan>=1.3.8
facexlib>=0.3.0

# ──────────────────────────────────────────────
# 얼굴 검출 fallback (HOG)
# ──────────────────────────────────────────────
dlib>=19.24.0
face-recognition>=1.3.0

# ──────────────────────────────────────────────
# 컬러화 (DDColor via modelscope)
# ──────────────────────────────────────────────
modelscope==1.36.2
datasets>=2.0.0
simplejson>=3.18.0
addict>=2.4.0
sortedcontainers>=2.4.0
oss2>=2.17.0
timm>=1.0.0

# ──────────────────────────────────────────────
# 이미지 분석 (BLIP, CLIP)
# ──────────────────────────────────────────────
transformers>=4.30.0
accelerate>=0.20.0

# ──────────────────────────────────────────────
# 정량 평가 (evaluation/ 모듈)
# ──────────────────────────────────────────────
scikit-image>=0.21.0   # PSNR, SSIM
lpips>=0.1.4            # LPIPS perceptual metric
piq>=0.7.0              # BRISQUE, NIQE 등 No-reference 지표
scipy>=1.10.0           # Wilcoxon signed-rank test
